+++ application/modules/journal/controllers/StatsController.php
@@ -39,16 +39,36 @@ class StatsController extends Zend_Controller_Action
             return;
         }
 
+
+        $journalSettings = Zend_Registry::get('reviewSettings');
+        $startStatsAfterDate = isset($journalSettings['startStatsAfterDate']) && $journalSettings['startStatsAfterDate'] !== '' ?
+            $journalSettings['startStatsAfterDate'] : null;
+
+
         /** @var Zend_Controller_Request_Http $request */
         $request = $this->getRequest();
         $yearQuery = (!empty($request->getParam('year'))) ? (int)$request->getParam('year') : null;
 
         $uri = 'journals/stats/dashboard/' . RVCODE;
 
+
         $errorMessage = "Une erreur s'est produite lors de la récupération des statistiques. Nous vous suggérons de ré-essayer dans quelques instants. Si le problème persiste vous devriez contacter le support de la revue.";
+        $params = ['withDetails' => ''];
+
+        if($yearQuery){
+            $params['year'] = $yearQuery;
+        }
+
+        $yearCategories = [];
+        $navYears = [];
+
+        if ($startStatsAfterDate) {
+            $params['startAfterDate'] = $startStatsAfterDate;
+            $this->view->startStatsAfterDate = $startStatsAfterDate;
+        }
 
         try { // api request
-            $dashboard = json_decode($this->askApi($uri, ['withDetails' => '', 'year' => $yearQuery]), true, 512, JSON_THROW_ON_ERROR);
+            $dashboard = json_decode($this->askApi($uri, $params), true, 512, JSON_THROW_ON_ERROR);
         } catch (GuzzleException $e) {
             $this->view->errorMessage = $errorMessage;
             trigger_error($e->getMessage());
@@ -61,13 +81,30 @@ class StatsController extends Zend_Controller_Action
         }
 
         $details = $dashboard['details'];
-        $yearCategories = array_keys($details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR]);
 
-        $this->view->yearCategories = $yearCategories; // navigation
+        if (isset($details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR])){
+            $yearCategories = array_keys($details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR]);
+        }
+
+
+
+       $navYears = $details[self::NB_SUBMISSIONS]['years']['indicator'];
+
+       if ($startStatsAfterDate){
+           $navYears = array_filter($navYears, static function($year) use($startStatsAfterDate){
+               return $year >= (int)date('Y', strtotime($startStatsAfterDate));
+           });
+
+       }
+
+
+        $this->view->yearCategories = $navYears; // navigation
 
         if ($yearQuery && !in_array($yearQuery, $yearCategories, true)) {
             Episciences_Tools::header('HTTP/1.1 404 Not Found');
-            $this->renderScript('index/notfound.phtml');
+            $this->view->message = $this->view->translate("Vous essayez de consulter les indicateurs statistiques pour l'année") . " <code>$yearQuery</code>";
+            $this->view->description = "Aucune information n'est disponible pour cette page pour le moment.";
+            $this->renderScript('error/error.phtml');
             return;
         }
 
@@ -75,6 +112,8 @@ class StatsController extends Zend_Controller_Action
         $series = [];
         $series[self::SUBMISSION_ACCEPTANCE_DELAY] = [];
         $series[self::SUBMISSION_PUBLICATION_DELAY] = [];
+        $series['submissionsByRepo'] = [];
+        $series[self::SUBMISSIONS_BY_YEAR] = [];
 
         $allPublications = $allRefusals = $allAcceptations = $allOtherStatus = 0;
         $publicationsPercentage = $acceptationsPercentage = $refusalsPercentage = $otherStatusPercentage = null;
@@ -83,8 +122,8 @@ class StatsController extends Zend_Controller_Action
             $yearCategories = [$yearQuery];
         }
 
-        $submissionsDelay = $dashboard['details']['averageDaysSubmissionAcceptation'];
-        $publicationsDelay = $dashboard['details']['averageDaysSubmissionPublication'];
+        $submissionsDelay = $details['averageDaysSubmissionAcceptation'];
+        $publicationsDelay = $details['averageDaysSubmissionPublication'];
         $allSubmissions = $dashboard['value'][self::NB_SUBMISSIONS]; // all review submissions
         $totalByYear = 0;
 
@@ -92,11 +131,12 @@ class StatsController extends Zend_Controller_Action
 
             $nbRefusals = $nbAcceptations = $nbOthers = 0;
 
-            $nbPublications = $details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['publications'];
+            $nbPublications = $details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['publications'] ?? 0;
             $allPublications += $nbPublications; // l'ensemble de la revue
 
             // stats collectées par rapport à la date de modification
-            $submissionsByYearResponse = array_key_exists($year, $details[self::NB_SUBMISSIONS][self::MORE_DETAILS]) ? $details[self::NB_SUBMISSIONS][self::MORE_DETAILS][$year] : [];
+            $moreDetails = $details[self::NB_SUBMISSIONS][self::MORE_DETAILS] ?? [];
+            $submissionsByYearResponse = array_key_exists($year, $moreDetails) ? $moreDetails[$year] : [];
 
             foreach ($submissionsByYearResponse as $values) {
 
@@ -134,12 +174,12 @@ class StatsController extends Zend_Controller_Action
 
             $totalByYear += $nbPublications;
 
-            $series[self::SUBMISSIONS_BY_YEAR]['submissions'][] = $dashboard['details'][self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['submissions']; // only submissions (1st version) of the current year
+            $series[self::SUBMISSIONS_BY_YEAR]['submissions'][] = $details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['submissions'] ?? 0; // only submissions (1st version) of the current year
             $series['acceptationByYear']['acceptations'][] = $nbAcceptations;
             $series['refusalsByYear']['refusals'][] = $nbRefusals;
             $series['publicationsByYear']['publications'][] = $nbPublications;
             $series['otherStatusByYear']['otherStatus'][] = $nbOthers; //totalNumberOfPapersAccepted
-            $series[self::SUBMISSIONS_BY_YEAR]['acceptedSubmittedSameYear'][] = $dashboard['details'][self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['acceptedSubmittedSameYear'];
+            $series[self::SUBMISSIONS_BY_YEAR]['acceptedSubmittedSameYear'][] = $details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['acceptedSubmittedSameYear'] ?? 0;
 
 
             if ($totalByYear) {
@@ -149,8 +189,10 @@ class StatsController extends Zend_Controller_Action
                 $series['otherStatusByYear']['percentage'][] = round($nbOthers / $totalByYear * 100, 2);
             }
 
+            $subByYear = $details[self::NB_SUBMISSIONS]['submissionsByRepo'][$year] ?? [];
+
             // submission by repo
-            foreach ($details[self::NB_SUBMISSIONS]['submissionsByRepo'][$year] as $repoId => $val) {
+            foreach ($subByYear as $repoId => $val) {
                 $series['submissionsByRepo'][$repoId][self::NB_SUBMISSIONS][] = $val['submissions'];
             }
 
@@ -173,7 +215,7 @@ class StatsController extends Zend_Controller_Action
 
         }
 
-        unset($nbPublications, $nbRefusals, $nbPublications, $nbOthers);
+        unset($nbPublications, $nbRefusals, $nbOthers);
 
         if ($yearQuery) {
             $allSubmissions = $series[self::SUBMISSIONS_BY_YEAR]['submissions'][0];
@@ -191,8 +233,8 @@ class StatsController extends Zend_Controller_Action
 
             unset($totalByYear);
 
-            $this->view->acceptedSubmittedSameYaer = $dashboard['details'][self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['acceptedSubmittedSameYear'];
-            $this->view->acceptationRateSubmittedSameYear = $dashboard['details'][self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['acceptanceRate'];
+            $this->view->acceptedSubmittedSameYaer = $details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['acceptedSubmittedSameYear'];
+            $this->view->acceptationRateSubmittedSameYear = $details[self::NB_SUBMISSIONS][self::SUBMISSIONS_BY_YEAR][$year]['acceptanceRate'];
 
 
         } elseif ($allSubmissions) {
@@ -225,13 +267,13 @@ class StatsController extends Zend_Controller_Action
             $this->view->translate("La répartition des <code>soumissions</code>par <code>année</code> et par <code>statut</code>") :
             $this->view->translate("La répartition des <code>soumissions</code> par <code>statut</code>");
 
-        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label1, 'data' => $series[self::SUBMISSIONS_BY_YEAR]['submissions'], 'backgroundColor' => self::COLORS_CODE[1]];
-        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label2, 'data' => $series['publicationsByYear']['publications'], 'backgroundColor' => self::COLORS_CODE[4]];
-        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label4, 'data' => $series['acceptationByYear']['acceptations'], 'backgroundColor' => self::COLORS_CODE[5]];
-        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label3, 'data' => $series['refusalsByYear']['refusals'], 'backgroundColor' => self::COLORS_CODE[2]];
+        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label1, 'data' => $series[self::SUBMISSIONS_BY_YEAR]['submissions'] ?? 0, 'backgroundColor' => self::COLORS_CODE[1]];
+        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label2, 'data' => $series['publicationsByYear']['publications'] ?? 0, 'backgroundColor' => self::COLORS_CODE[4]];
+        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label4, 'data' => $series['acceptationByYear']['acceptations'] ?? 0, 'backgroundColor' => self::COLORS_CODE[5]];
+        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label3, 'data' => $series['refusalsByYear']['refusals'] ?? 0, 'backgroundColor' => self::COLORS_CODE[2]];
 
-        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label5, 'data' => $series['otherStatusByYear']['otherStatus'], 'backgroundColor' => self::COLORS_CODE[0]];
-        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label6, 'data' => $series[self::SUBMISSIONS_BY_YEAR]['acceptedSubmittedSameYear'], 'backgroundColor' => self::COLORS_CODE[6]];
+        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label5, 'data' => $series['otherStatusByYear']['otherStatus'] ?? 0, 'backgroundColor' => self::COLORS_CODE[0]];
+        $seriesJs[self::SUBMISSIONS_BY_YEAR]['datasets'][] = ['label' => $label6, 'data' => $series[self::SUBMISSIONS_BY_YEAR]['acceptedSubmittedSameYear'] ?? 0, 'backgroundColor' => self::COLORS_CODE[6]];
 
 
         $seriesJs[self::SUBMISSIONS_BY_YEAR]['chartType'] = self::CHART_TYPE['BAR'];
@@ -269,7 +311,7 @@ class StatsController extends Zend_Controller_Action
 
         if ($isAvailableUsersStats) {
             $allUsers = $dashboard['value']['nbUsers'];
-            $usersDetails = $dashboard['details']['nbUsers'];
+            $usersDetails = $details['nbUsers'];
             $roles = array_keys($usersDetails);
             $rootKey = array_search(Episciences_Acl::ROLE_ROOT, $roles, true);
 
@@ -296,6 +338,31 @@ class StatsController extends Zend_Controller_Action
 
 
         $this->view->allSubmissionsJs = $allSubmissions;
+
+        if (!$yearQuery) {
+
+            try {
+                $totalPublishedArticles = (int)json_decode($this->askApi('journals/stats/nb-submissions/' . RVCODE, ['status' => Episciences_Paper::STATUS_PUBLISHED]), true, 512, JSON_THROW_ON_ERROR)['value'];
+                $totalArticles = (int)json_decode($this->askApi('journals/stats/nb-submissions/' . RVCODE), true, 512, JSON_THROW_ON_ERROR)['value'];
+                $totalImportedArticles = (int)json_decode($this->askApi('journals/stats/nb-submissions/' . RVCODE, ['flag' => 'imported']), true, 512, JSON_THROW_ON_ERROR)['value'];
+
+                if ($totalImportedArticles > 0) {
+                    $this->view->totalImportedArticles = $totalImportedArticles;
+                }
+
+                if ($totalPublishedArticles > 0) {
+                    $this->view->totalPublishedArticles = $totalPublishedArticles;
+                }
+
+                $this->view->totalArticles = $totalArticles;
+            } catch (GuzzleException|JsonException  $e) {
+                trigger_error($e->getMessage());
+
+            }
+
+
+        }
+
         $this->view->allPublications = !$yearQuery ? $dashboard['value']['totalPublished'] : $allPublications;
         $this->view->allRefusals = $allRefusals;
         $this->view->allAcceptations = $allAcceptations;